🚚 (consent-api): Move create services to create.ts
