+++ apps/consent-api/src/routers/participantResponses.ts
@@ -19,7 +19,7 @@
 
 import { Router } from 'express';
 
-import { createResponse } from '../service/participantResponses.js';
+import { createResponse } from '../service/create.js';
 
 /**
  * @openapi
